Feature: Halloween Bot Scare
The bot reacts to the command boo!booh and joins your current voice
channel and plays randomly a sound to scare you and your companions in
the channel. A real laughter for everyone who is present in the channel!

There is for the MVP Feature no spam protection nor queuing implemented.